requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
diffusers==0.21.4
transformers==4.33.1
torch==2.0.1
peft==0.5.0
Pillow==10.0.0
python-multipart==0.0.6
gunicorn==21.2.0
